Reorganized sub-classes to make them more visible.
