+++ TravelTimes/src/gov/usgs/locator/AuxTtRef.java
@@ -3,8 +3,10 @@ package gov.usgs.locator;
 import java.io.BufferedInputStream;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.RandomAccessFile;
+// import java.io.RandomAccessFile;
 import java.util.ArrayList;
+import java.util.Map;
+import java.util.NavigableMap;
 import java.util.Scanner;
 import java.util.TreeMap;
 
@@ -22,20 +24,28 @@ import java.util.TreeMap;
  *
  */
 public class AuxTtRef {
-	final PhGroup regional;								// Regional phase group
-	final PhGroup depth;									// Depth sensitive phase group
-	final PhGroup downWeight;							// Phases to be down weighted
-	final ArrayList<PhGroup> phGroups;		// List of primary phase groups
-	final ArrayList<PhGroup> auxGroups;		// List of auxiliary phase groups
-	final TreeMap<String, TtStat> ttStats;			// List of phase statistics
+	// Phase group storage.
+	final PhGroup regional;									// Regional phase group
+	final PhGroup depth;										// Depth sensitive phase group
+	final PhGroup downWeight;								// Phases to be down weighted
+	final PhGroup canUse;										// Phases that can be used for location
+	final ArrayList<PhGroup> phGroups;			// List of primary phase groups
+	final ArrayList<PhGroup> auxGroups;			// List of auxiliary phase groups
+	// Phase statistics storage.
+	final TreeMap<String, TtStat> ttStats;	// List of phase statistics
+	// Ellipticity storage.
+	final TreeMap<String, Ellip> ellips;		// List of ellipticity corrections
+	// Flag storage by phase
+	final TreeMap<String, TtFlags> phFlags;		// Phase group information by phase
+	boolean hasBounce;			// True if the phase bounces from the free surface
+	TtStat ttStat;					// Phase statistics
+	Ellip ellip;						// Ellipticity correction
 	// Set up the reader.
 	final String phGroupPath = "../../../Documents/Work/Models/phgrp.dat";
 	final String ttStatsPath = "../../../Documents/Work/Models/ttstats.lis";
-	RandomAccessFile inGroup;
-	BufferedInputStream inStats;
+	final String ellipPath = "../../../Documents/Work/Models/tau.table";
 	Scanner scan;
-	byte[] data;
-	int cur = 0;
+	int nDepth;
 	String nextCode;
 
 	/**
@@ -46,44 +56,53 @@ public class AuxTtRef {
 	 * Locator.
 	 * 
 	 * @param printGrp If true, print the phase groups
+	 * @param printFlg If true, print the phase flags
 	 * @param printRaw If true, print the raw statistics data
 	 * @param printFit If true, print the fit statistics data
 	 * @throws IOException If opens fail
 	 */
-	public AuxTtRef(boolean printGrp, boolean printRaw, boolean printFit) 
+	public AuxTtRef(boolean printGrp, boolean printFlg, boolean printRaw, boolean printFit) 
 			throws IOException {
+		BufferedInputStream inGroup, inStats, inEllip;
 		TtStat ttStat;
+		Ellip ellip;
+		EllipDeps eDepth;
 		
 		// Open and read the phase groups file.
-		inGroup = new RandomAccessFile(phGroupPath, "r");
-		data = new byte[(int)inGroup.length()];
-		inGroup.read(data);
+		inGroup = new BufferedInputStream(new FileInputStream(phGroupPath));
+		scan = new Scanner(inGroup);
+		// Prime the pump.
+		nextCode = scan.next();
 		// Handle local-regional phases separately.
 		regional = read1Group();
 		if(printGrp) {
 			regional.dumpGroup();
 			System.out.println();
 		}
-		cur += 2;
 		// Handle depth phases separately.
 		depth = read1Group();
 		if(printGrp) {
 			depth.dumpGroup();
 			System.out.println();
 		}
-		cur += 2;
 		// Handle down weighted phases separately.
 		downWeight = read1Group();
 		if(printGrp) {
 			downWeight.dumpGroup();
 			System.out.println();
 		}
-		cur += 2;
+		// Handle used phases separately.
+		canUse = read1Group();
+		if(printGrp) {
+			canUse.dumpGroup();
+			System.out.println();
+		}
 		
-		// Handle everything else.
+		// Handle "normal" groups.
 		phGroups = new ArrayList<PhGroup>();
 		auxGroups = new ArrayList<PhGroup>();
 		readGroups(printGrp);
+		inGroup.close();
 		
 		// Open and read the travel-time statistics file.
 		inStats = new BufferedInputStream(new FileInputStream(ttStatsPath));
@@ -95,8 +114,26 @@ public class AuxTtRef {
 		do {
 			ttStat = read1StatHead();
 			ttStats.put(ttStat.phCode, ttStat);
-			read1StatData(new LinearFit(ttStat), printRaw, printFit);
-		} while(!statEof());
+			read1StatData(new TtStatLinFit(ttStat), printRaw, printFit);
+		} while(scan.hasNext());
+		inStats.close();
+		
+		// Open and read the ellipticity correction file.
+		inEllip = new BufferedInputStream(new FileInputStream(ellipPath));
+		scan = new Scanner(inEllip);
+		ellips = new TreeMap<String, Ellip>();
+		eDepth = new EllipDeps();
+		nDepth = eDepth.ellipDeps.length;
+		do {
+			ellip = read1Ellip();
+			ellips.put(ellip.phCode, ellip);
+		} while(scan.hasNext());
+		inEllip.close();
+		
+		// Rearrange group flags, phase flags and statistics and the 
+		// ellipticity correction by phase.
+		phFlags = new TreeMap<String, TtFlags>();
+		makePhFlags(printFlg);
 	}
 	
 	/**
@@ -104,52 +141,20 @@ public class AuxTtRef {
 	 * 
 	 * @return Phase group just read
 	 */
-	protected PhGroup read1Group() {
-		char[] buffer = new char[8];
-		int ind;
-		boolean use;
-		
-		// Check to see if we have a blank line.
-		if(cur >= data.length) return null;
-		if((char)data[cur] == '\r') {
-			cur += 2;
+	private PhGroup read1Group() {
+		if(nextCode.contains(":")) {
+			PhGroup group = new PhGroup(nextCode.substring(0, 
+					nextCode.indexOf(':')));
+			nextCode = scan.next();
+			while(!nextCode.contains(":") & !nextCode.contains("-")) {
+				group.addPhase(nextCode);
+				nextCode = scan.next();
+			}
+			return group;
+		} else {
+			if(scan.hasNext()) nextCode = scan.next();
 			return null;
 		}
-		
-		// First get the group name.
-		ind = 0;
-		for(int j=0; j<buffer.length; j++) buffer[j] = ' ';
-		// Start by skipping leading blanks.
-		while((char)data[cur] == ' ') cur++;
-		// Look for the next delimiter.
-		while((char)data[cur] != '*' && (char)data[cur] != ':') 
-			buffer[ind++] = (char)data[cur++];
-		// Check for the use flag.
-		if((char)data[cur] == '*') {
-			use = true;
-			while((char)data[cur] != ':') cur++;
-		}
-		else use = false;
-		// Create the group.
-		PhGroup group = new PhGroup(new String(buffer), use);
-		
-		// Collect the phases in the group.
-		do  {
-			cur++;
-			ind = 0;
-			for(int j=0; j<buffer.length; j++) buffer[j] = ' ';
-			// Skip leading blanks.
-			while((char)data[cur] == ' ') cur++;
-			// Look for the next delimiter.
-			while((char)data[cur] != ',' && (char)data[cur] != '\r') 
-				buffer[ind++] = (char)data[cur++];
-			// Add the phase.
-			group.addPhase(new String(buffer));
-		} while((char)data[cur] != '\r');
-		
-		// Done with this line.
-		cur += 2;
-		return group;
 	}
 	
 	/**
@@ -160,7 +165,7 @@ public class AuxTtRef {
 	 * 
 	 * @param print List the auxiliary data as it's read if true
 	 */
-	protected void readGroups(boolean print) {
+	private void readGroups(boolean print) {
 		do {
 			// Groups are added to the ArrayLists as they are created.
 			phGroups.add(read1Group());
@@ -171,7 +176,7 @@ public class AuxTtRef {
 					auxGroups.get(auxGroups.size()-1).dumpGroup();
 				else System.out.println("    *null*");
 			}
-		} while(cur < data.length);
+		} while(scan.hasNext());
 	}
 	
 	/**
@@ -232,15 +237,10 @@ public class AuxTtRef {
 	 * @param groupName Phase group name
 	 * @return True if this group can be used for earthquake location
 	 */
-	public boolean useGroup(String groupName) {
-		for(int j=0; j<phGroups.size(); j++) {
-			if(groupName.equals(phGroups.get(j).groupName)) {
-				return phGroups.get(j).useInLoc;
-			}
-			if(auxGroups.get(j) != null) {
-				if(groupName.equals(auxGroups.get(j).groupName)) {
-					return auxGroups.get(j).useInLoc;
-				}
+	public boolean canUse(String phase) {
+		for(int k=0; k<canUse.phases.size(); k++) {
+			if(phase.equals(canUse.phases.get(k))) {
+				return true;
 			}
 		}
 		return false;
@@ -296,7 +296,7 @@ public class AuxTtRef {
 	 * 
 	 * @return Statistics object
 	 */
-	protected TtStat read1StatHead() {
+	private TtStat read1StatHead() {
 		String phCode;
 		int minDelta, maxDelta;
 		TtStat ttStat;
@@ -320,7 +320,7 @@ public class AuxTtRef {
 	 * @param printRaw If true, print the raw statistics data
 	 * @param printFit If true, print the fit statistics data
 	 */
-	protected void read1StatData(LinearFit fit, boolean printRaw, 
+	private void read1StatData(TtStatLinFit fit, boolean printRaw, 
 			boolean printFit) {
 		int delta;
 		double res, spd, obs;
@@ -391,17 +391,6 @@ public class AuxTtRef {
 	}
 	
 	/**
-	 * Check the statistics for an end-of-file.
-	 * 
-	 * @return True if all the data has been read.
-	 */
-	protected boolean statEof() {
-		
-		if(scan.hasNext()) return false;
-		else return true;
-	}
-	
-	/**
 	 * Find the statistics associated with the desired phase.
 	 * 
 	 * @param phase Phase code
@@ -419,17 +408,8 @@ public class AuxTtRef {
 	 * @return Bias in seconds at distance delta
 	 */
 	public double getBias(TtStat ttStat, double delta) {
-		StatSeg seg;
-		
-		if(ttStat == null) return 0d;
-		
-		for(int k=0; k<ttStat.bias.size(); k++) {
-			seg = ttStat.bias.get(k);
-			if(delta >= seg.minDelta && delta <= seg.maxDelta) {
-				return seg.interp(delta);
-			}
-		}
-		return 0d;
+		if(ttStat == null) return TauUtil.DEFBIAS;
+		else return ttStat.getBias(delta);
 	}
 	
 	/**
@@ -440,17 +420,8 @@ public class AuxTtRef {
 	 * @return Spread in seconds at distance delta
 	 */
 	public double getSpread(TtStat ttStat, double delta) {
-		StatSeg seg;
-		
-		if(ttStat == null) return 12d;
-		
-		for(int k=0; k<ttStat.spread.size(); k++) {
-			seg = ttStat.spread.get(k);
-			if(delta >= seg.minDelta && delta <= seg.maxDelta) {
-				return seg.interp(delta);
-			}
-		}
-		return 12d;
+		if(ttStat == null) return TauUtil.DEFSPREAD;
+		else return ttStat.getSpread(delta);
 	}
 	
 	/**
@@ -461,358 +432,150 @@ public class AuxTtRef {
 	 * @return Relative observability at distance delta
 	 */
 	public double getObserv(TtStat ttStat, double delta) {
-		StatSeg seg;
-		
-		if(ttStat == null) return 0d;
-		
-		for(int k=0; k<ttStat.observ.size(); k++) {
-			seg = ttStat.observ.get(k);
-			if(delta >= seg.minDelta && delta <= seg.maxDelta) {
-				return seg.interp(delta);
-			}
-		}
-		return 0d;
+		if(ttStat == null) return TauUtil.DEFOBSERV;
+		else return ttStat.getObserv(delta);
 	}
-}
-
-/**
- * Helper class holding one phase group.
- * 
- * @author Ray Buland
- *
- */
-class PhGroup {
-	String groupName;							// Name of the phase group
-	ArrayList<String> phases;			// List of phases in the group
-	boolean useInLoc;							// True if can be used in a location
 	
 	/**
-	 * Initialize the phase group.
+	 * Read in ellipticity correction data for one phase.
 	 * 
-	 * @param groupName Name of the phase group
-	 * @param useInLoc May be used in an earthquale location if true
+	 * @return Ellipticity object
 	 */
-	protected PhGroup(String groupName, boolean useInLoc) {
-		this.groupName = groupName.trim();
-		this.useInLoc = useInLoc;
-		phases = new ArrayList<String>();
-	}
-	
-	/**
-	 * Add one phase to the group.
-	 * 
-	 * @param phase Phase code to be added
-	 */
-	protected void addPhase(String phase) {
-		phases.add(phase.trim());
-	}
-	
-	/**
-	 * Print the contents of this phase group.
-	 */
-	protected void dumpGroup() {
-		System.out.print(groupName+": ");
-		for(int j=0; j<phases.size(); j++) {
-			if(j <= 0) System.out.print(phases.get(j));
-			else System.out.print(" ,"+phases.get(j));
+	private Ellip read1Ellip() {
+		String phCode;
+		int nDelta;
+		@SuppressWarnings("unused")
+		double delta;
+		double minDelta, maxDelta;
+		double[][] t0, t1, t2;
+		Ellip ellip;
+		
+		// Read the header.
+		phCode = scan.next();
+		nDelta = scan.nextInt();
+		minDelta = scan.nextDouble();
+		maxDelta = scan.nextDouble();
+		
+		// Allocate storage.
+		t0 = new double[nDelta][nDepth];
+		t1 = new double[nDelta][nDepth];
+		t2 = new double[nDelta][nDepth];
+		// Read in the tau profiles.
+		for(int j=0; j<nDelta; j++) {
+			delta = scan.nextDouble();		// The distance is cosmetic
+			for(int i=0; i<nDepth; i++) t0[j][i] = scan.nextDouble();
+			for(int i=0; i<nDepth; i++) t1[j][i] = scan.nextDouble();
+			for(int i=0; i<nDepth; i++) t2[j][i] = scan.nextDouble();
 		}
-		System.out.println(" "+useInLoc);
+		// Return the result.
+		ellip = new Ellip(phCode, minDelta, maxDelta, t0, t1, t2);
+		return ellip;
 	}
-}
-
-/**
- * Helper class holding one set of phase statistics.
- * 
- * @author Ray Buland
- *
- */
-class TtStat {
-	String phCode;							// Phase code
-	int minDelta;								// Minimum distance in degrees
-	int maxDelta;								// Maximum distance in degrees
-	ArrayList<StatSeg> bias;		// Measured arrival time bias (s)
-	ArrayList<StatSeg> spread;	// Measured residual spread (s)
-	ArrayList<StatSeg> observ;	// Measured observability
 	
 	/**
-	 * Initialize the phase statistics.
+	 * Get the ellipticity correction data for the desired phase.
 	 * 
 	 * @param phCode Phase code
-	 * @param minDelta Minimum observed distance in degrees
-	 * @param maxDelta Maximum observed distance in degrees
+	 * @return Ellipticity data
 	 */
-	protected TtStat(String phCode, int minDelta, int maxDelta) {
-		this.phCode = phCode;
-		this.minDelta = minDelta;
-		this.maxDelta = maxDelta;
-		// set up storage for the linear fits.
-		bias = new ArrayList<StatSeg>();
-		spread = new ArrayList<StatSeg>();
-		observ = new ArrayList<StatSeg>();
+	public Ellip findEllip(String phCode) {
+		return ellips.get(phCode);
 	}
 	
 	/**
-	 * Print the travel-time statistics.
+	 * Reorganize the flags from ArrayLists of phases by group to 
+	 * a TreeMap of flags by phase.
 	 */
-	protected void dumpStats() {
-		// Print the header.
-		System.out.println("\n"+phCode+"     "+minDelta+"     "+maxDelta);
-		
-		// Print the data.
-		System.out.println("Bias:");
-		for(int j=0; j<bias.size(); j++) {
-			System.out.format("  %3d  range = %6.2f, %6.2f  fit = %11.4e, "+
-					"%11.4e\n",j,bias.get(j).minDelta,bias.get(j).maxDelta,
-					bias.get(j).slope,bias.get(j).offset);
-		}
-		System.out.println("Spread:");
-		for(int j=0; j<spread.size(); j++) {
-			System.out.format("  %3d  range = %6.2f, %6.2f  fit = %11.4e, "+
-					"%11.4e\n",j,spread.get(j).minDelta,spread.get(j).maxDelta,
-					spread.get(j).slope,spread.get(j).offset);
-		}
-		System.out.println("Observability:");
-		for(int j=0; j<observ.size(); j++) {
-			System.out.format("  %3d  range = %6.2f, %6.2f  fit = %11.4e, "+
-					"%11.4e\n",j,observ.get(j).minDelta,observ.get(j).maxDelta,
-					observ.get(j).slope,observ.get(j).offset);
-		}
-	}
-}
+	private void makePhFlags(boolean print) {
+		String phCode, phGroup;
+		TtFlags flags;
 
-/**
- * Helper class holding one linear interpolation segment.
- * 
- * @author Ray Buland
- *
- */
-class StatSeg {
-	double minDelta;						// Minimum distance in degrees
-	double maxDelta;						// Maximum distance in degrees
-	double slope;								// Slope of linear interpolation
-	double offset;							// Offset of linear interpolation
-	
-	/**
-	 * Create the linear segment.
-	 * 
-	 * @param minDelta Minimum distance in degrees
-	 * @param maxDelta Maximum distance in degrees
-	 * @param slope Slope of the linear fit
-	 * @param offset Offset of the linear fit
-	 */
-	protected StatSeg(double minDelta, double maxDelta, 
-			double slope, double offset) {
-		this.minDelta = minDelta;
-		this.maxDelta = maxDelta;
-		this.slope = slope;
-		this.offset = offset;
-	}
-	
-	/**
-	 * Interpolate the linear fit at one distance.
-	 * 
-	 * @param delta Distance in degrees where statistics are desired
-	 * @return Interpolated parameter
-	 */
-	protected double interp(double delta) {
-		if(delta >= minDelta && delta <= maxDelta) {
-			return offset+delta*slope;
-		} else return(Double.NaN);
-	}
-}
-
-/**
- * Helper class to acquire the 1 degree statistics data and do 
- * the linear fits.
- * 
- * @author Ray Buland
- *
- */
-class LinearFit {
-	double[] res, spd, obs;
-	boolean[] resBrk, spdBrk, obsBrk;
-	TtStat ttStat;
-	int minDelta, maxDelta;
-	
-	/**
-	 * Set up the 1 degree arrays.
-	 * 
-	 * @param ttStat Phase statistics object
-	 */
-	protected LinearFit(TtStat ttStat) {
-		this.ttStat = ttStat;
-		minDelta = ttStat.minDelta;
-		maxDelta = ttStat.maxDelta;
-		
-		// set up the 1 degree statistics arrays.
-		res = new double[maxDelta-minDelta+1];
-		resBrk = new boolean[maxDelta-minDelta+1];
-		spd = new double[maxDelta-minDelta+1];
-		spdBrk = new boolean[maxDelta-minDelta+1];
-		obs = new double[maxDelta-minDelta+1];
-		obsBrk = new boolean[maxDelta-minDelta+1];
-		// Initialize the arrays.
-		for(int j=0; j<res.length; j++) {
-			res[j] = Double.NaN;
-			resBrk[j] = false;
-			spd[j] = Double.NaN;
-			spdBrk[j] = false;
-			obs[j] = Double.NaN;
-			obsBrk[j] = false;
-		}
-	}
-	
-	/**
-	 * Add statistics for one 1 degree distance bin.
-	 * 
-	 * @param delta Distance in degrees at the bin center
-	 * @param res Travel-time residual bias in seconds
-	 * @param resBrk Break the interpolation at this bin if true
-	 * @param spd Robust estimate of the scatter of travel-time 
-	 * residuals
-	 * @param spdBrk Break the interpolation at this bin if true
-	 * @param obs Number of times this phase was observed in the 
-	 * defining study
-	 * @param obsBrk Break the interpolation at this bin if true
-	 */
-	protected void add(int delta, double res, boolean resBrk, 
-			double spd, boolean spdBrk, double obs, boolean obsBrk) {
-		this.res[delta-minDelta] = res;
-		this.resBrk[delta-minDelta] = resBrk;
-		this.spd[delta-minDelta] = spd;
-		this.spdBrk[delta-minDelta] = spdBrk;
-		this.obs[delta-minDelta] = obs;
-		this.obsBrk[delta-minDelta] = obsBrk;
-	}
-	
-	/**
-	 * Do the linear fits for all statistics variables.
-	 */
-	protected void fitAll() {
-		doFits(ttStat.bias, res, resBrk);
-		doFits(ttStat.spread, spd, spdBrk);
-		doFits(ttStat.observ, obs, obsBrk);
-	}
-	
-	/**
-	 * Do the linear fits for all segments of one statistics variable.
-	 * 
-	 * @param interp The ArrayList where the fits will be stored
-	 * @param value Array of parameter values
-	 * @param brk Array of break point flags
-	 */
-	protected void doFits(ArrayList<StatSeg> interp, double[] value, 
-			boolean[] brk) {
-		int start, end = 0;
-		double startDelta, endDelta;
-		double[] b;
-		
-		// Find the break points and invoke the fitter for each segment.
-		endDelta = minDelta;
-		for(int j=0; j<value.length; j++) {
-			if(brk[j]) {
-				start = end;
-				end = j;
-				startDelta = endDelta;
-				endDelta = minDelta+(double)j;
-				b = do1Fit(start, end, minDelta, value);
-				interp.add(new StatSeg(startDelta, endDelta, b[0], b[1]));
+		// Search the phase groups for phases.
+		for(int j=0; j<phGroups.size(); j++) {
+			phGroup = phGroups.get(j).groupName;
+			for(int i=0; i<phGroups.get(j).phases.size(); i++) {
+				phCode = phGroups.get(j).phases.get(i);
+				unTangle(phCode, phGroup);
+				phFlags.put(phCode, new TtFlags(phGroup, compGroup(phGroup), 
+						isRegional(phCode), isDepthPh(phCode), canUse(phCode), 
+						isDisPh(phCode), hasBounce, ttStat, ellip));
 			}
 		}
-		// Fit the last segment.
-		start = end;
-		end = value.length-1;
-		startDelta = endDelta;
-		endDelta = minDelta+(double)end;
-		b = do1Fit(start, end, minDelta, value);
-		interp.add(new StatSeg(startDelta, endDelta, b[0], b[1]));
-		fixEnds(interp);
-	}
-	
-	/**
-	 * Do the linear fit for one segment of one statistics variable.
-	 * 
-	 * @param start Array index of the start of the segment
-	 * @param end Array index of the end of the segment
-	 * @param minDelta Minimum distance where the phase is observed 
-	 * in degrees
-	 * @param value Raw statistics data to be fit
-	 * @return An array containing the fit slope and offset
-	 */
-	protected double[] do1Fit(int start, int end, double minDelta, 
-			double[] value) {
-		double[][] a = new double[2][2];
-		double[] y = new double[2];
-		double[] b = new double[2];
-		double delta, det;
-		
-		// Initialize temporary storage.
-		for(int i=0; i<2; i++) {
-			y[i] = 0d;
-			for(int j=0; j<2; j++) {
-				a[i][j] = 0d;
+		// Search the auxiliary phase groups for phases.
+		for(int j=0; j<auxGroups.size(); j++) {
+			if(auxGroups.get(j) != null) {
+				phGroup = auxGroups.get(j).groupName;
+				for(int i=0; i<auxGroups.get(j).phases.size(); i++) {
+					phCode = auxGroups.get(j).phases.get(i);
+					unTangle(phCode, phGroup);
+					phFlags.put(phCode, new TtFlags(phGroup, compGroup(phGroup), 
+							isRegional(phCode), isDepthPh(phCode), canUse(phCode), 
+							isDisPh(phCode), hasBounce, ttStat, ellip));
+				}
 			}
 		}
 		
-		// Skip null bins and collect the data available.
-		for(int j=start; j<=end; j++) {
-			if(!Double.isNaN(value[j])) {
-				delta = minDelta+j;
-				y[0] += value[j]*delta;
-				y[1] += value[j];
-				a[0][0] += 1d;
-				a[0][1] -= delta;
-				a[1][1] += Math.pow(delta, 2);
+		if(print) {
+			NavigableMap<String, TtFlags> map = phFlags.headMap("~", true);
+			System.out.println("\n     Phase Flags:");
+			for(@SuppressWarnings("rawtypes") Map.Entry entry : map.entrySet()) {
+				flags = (TtFlags)entry.getValue();
+				System.out.format("%8s: %8s %8s  flags = %5b %5b %5b %5b %5b", 
+						entry.getKey(), flags.phGroup, flags.auxGroup, flags.canUse, 
+						flags.isRegional, flags.isDepth, flags.dis, flags.hasBounce);
+				if(flags.ttStat == null) System.out.print(" stats = null    ");
+				else System.out.format(" stats = %-8s", flags.ttStat.phCode);
+				if(flags.ellip == null) System.out.println(" ellip = null");
+				else System.out.format(" ellip = %s\n", flags.ellip.phCode);
 			}
 		}
-		a[1][0] = a[0][1];
-		
-		// Do the fit.
-		det = a[0][0]*a[1][1]-a[0][1]*a[1][0];
-		b[0] = (a[0][0]*y[0]+a[0][1]*y[1])/det;
-		b[1] = (a[1][0]*y[0]+a[1][1]*y[1])/det;
-		return b;
 	}
 	
 	/**
-	 * Successive linear fits don't quite connect with each other at 
-	 * exactly the break point distances.  Compute the actual cross-
-	 * over distances and apply them to the end of one segment and 
-	 * the start of the next.
+	 * Do some fiddling to determine if the phase bounces from the 
+	 * free surface and to add the statistics and ellipticity 
+	 * correction.
 	 * 
-	 * @param interp ArrayList of linear fit segments for one parameter
+	 * @param phCode Phase code
+	 * @param phGroup Group code
 	 */
-	protected void fixEnds(ArrayList<StatSeg> interp) {
-		StatSeg last, cur;
-		
-		cur = interp.get(0);
-		for(int j=1; j<interp.size(); j++) {
-			last = cur;
-			cur = interp.get(j);
-			last.maxDelta = -(last.offset-cur.offset)/(last.slope-cur.slope);
-			cur.minDelta = last.maxDelta;
+	private void unTangle(String phCode, String phGroup) {		
+		// See if the phase bounces at the free surface.
+		hasBounce = false;
+		if(phCode.charAt(0) == 'p' || phCode.charAt(0) == 's') {
+			hasBounce = true;
+		} else if(phCode.equals("pwP")) {
+			hasBounce = true;
+		} else if(phGroup.equals("P'P'") || phGroup.equals("S'S'")) {
+			hasBounce = true;
+		} else {
+			for(int j=0; j<phGroup.length()-1; j++) {
+				if((phGroup.charAt(j) == 'P' || phGroup.charAt(j) == 'S') && 
+						(phGroup.charAt(j+1) == 'P' || phGroup.charAt(j+1) == 'S')) {
+					hasBounce = true;
+					break;
+				}
+			}
+		}
+		// Get the travel-time statistics.
+		ttStat = findStats(phCode);
+		// The ellipticity is a little messier.
+		ellip = findEllip(phCode);
+		if(ellip == null) ellip = findEllip(phGroup);
+		if(ellip == null) {
+			if(phCode.equals("pwP")) ellip = findEllip("pP");
+			else if(phCode.equals("PKPpre")) ellip = findEllip("PKPdf");
+			else if(phGroup.contains("PKP")) ellip = findEllip(phGroup+"bc");
 		}
 	}
 	
 	/**
-	 * Print the travel-time statistics.
+	 * Get flags, etc. by phase code.
+	 * 
+	 * @param phCode Phase code
+	 * @return Flags object
 	 */
-	protected void dumpStats() {
-		char[] flag;
-		// Print the header.
-		System.out.println("\n"+ttStat.phCode+"     "+minDelta+"     "+maxDelta);
-		
-		// If the arrays still exist, dump the raw statistics.
-		flag = new char[3];
-		for(int j=0; j<res.length; j++) {
-			if(resBrk[j]) flag[0] = '*';
-			else flag[0] = ' ';
-			if(spdBrk[j]) flag[1] = '*';
-			else flag[1] = ' ';
-			if(obsBrk[j]) flag[2] = '*';
-			else flag[2] = ' ';
-			System.out.format("  %3d  %7.2f%c  %7.2f%c  %8.1f%c\n",j+minDelta,
-					res[j],flag[0],spd[j],flag[1],obs[j],flag[2]);
-		}
+	public TtFlags findFlags(String phCode) {
+		return phFlags.get(phCode);
 	}
 }
